ProofOfConcept
A hand-developed example of providing both component-independent and
component-dependent interfaces for a class that may be shared by
different components.

This work is driven by the desire to to pass instances as inst_ref
parameters in port messages.

The hand-modified code is derived from compiled code and stashed in
src/main/java to verify compilation.